fix: Relax battery communication timings
Battery communication timing was quite short, sometimes the battery
doesn't react/respond in time and we give up too early. Especially
when first opening the seatbox after the batteries were asleep, there
appears to be a delay in activation.

This relaxes the timings, which should be counteracted by the new more
aggressive retry logic.